move error handling back to client package
