+++ client/burn_alert_test.go
@@ -9,7 +9,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/honeycombio/terraform-provider-honeycombio/client"
-	"github.com/honeycombio/terraform-provider-honeycombio/client/errors"
 	"github.com/honeycombio/terraform-provider-honeycombio/internal/helper/test"
 )
 
@@ -220,7 +219,7 @@ func TestBurnAlerts(t *testing.T) {
 		t.Run(fmt.Sprintf("Fail to GET a deleted burn alert: %s", testName), func(t *testing.T) {
 			_, err := c.BurnAlerts.Get(ctx, dataset, burnAlert.ID)
 
-			var de errors.DetailedError
+			var de client.DetailedError
 			require.Error(t, err)
 			require.ErrorAs(t, err, &de)
 			assert.True(t, de.IsNotFound())
