+++ client/client.go
@@ -19,8 +19,6 @@ import (
 
 	cleanhttp "github.com/hashicorp/go-cleanhttp"
 	retryablehttp "github.com/hashicorp/go-retryablehttp"
-
-	hnyerr "github.com/honeycombio/terraform-provider-honeycombio/client/errors"
 )
 
 const (
@@ -223,7 +221,7 @@ func (c *Client) Do(ctx context.Context, method, path string, requestBody, respo
 	defer resp.Body.Close()
 
 	if !(resp.StatusCode >= 200 && resp.StatusCode <= 299) {
-		return hnyerr.FromResponse(resp)
+		return FromResponse(resp)
 	}
 	if responseBody != nil {
 		err = json.NewDecoder(resp.Body).Decode(responseBody)
