+++ client/marker.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"net/http"
 	"time"
-
-	"github.com/honeycombio/terraform-provider-honeycombio/client/errors"
 )
 
 // Markers describes all the marker-related methods that the Honeycomb API
@@ -90,7 +88,7 @@ func (s *markers) Get(ctx context.Context, dataset string, id string) (*Marker, 
 			return &m, nil
 		}
 	}
-	return nil, errors.DetailedError{
+	return nil, DetailedError{
 		Status:  http.StatusNotFound,
 		Message: "Marker Not Found.",
 	}
