+++ client/marker_settings.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"net/http"
 	"time"
-
-	"github.com/honeycombio/terraform-provider-honeycombio/client/errors"
 )
 
 // MarkerSettings describes all the markerType-related methods that the Honeycomb API
@@ -77,7 +75,7 @@ func (s *markerSettings) Get(ctx context.Context, dataset string, id string) (*M
 			return &m, nil
 		}
 	}
-	return nil, errors.DetailedError{
+	return nil, DetailedError{
 		Status:  http.StatusNotFound,
 		Message: "Marker Setting Not Found.",
 	}
