+++ client/marker_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/honeycombio/terraform-provider-honeycombio/client"
-	"github.com/honeycombio/terraform-provider-honeycombio/client/errors"
 	"github.com/honeycombio/terraform-provider-honeycombio/internal/helper/test"
 )
 
@@ -95,7 +94,7 @@ func TestMarkers(t *testing.T) {
 	t.Run("Fail to Get deleted Marker", func(t *testing.T) {
 		_, err := c.Markers.Get(ctx, dataset, m.ID)
 
-		var de errors.DetailedError
+		var de client.DetailedError
 		require.Error(t, err)
 		require.ErrorAs(t, err, &de)
 		assert.True(t, de.IsNotFound())
