+++ honeycombio/resource_marker_setting.go
@@ -11,7 +11,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
 	honeycombio "github.com/honeycombio/terraform-provider-honeycombio/client"
-	hnyerr "github.com/honeycombio/terraform-provider-honeycombio/client/errors"
 )
 
 func newMarkerSetting() *schema.Resource {
@@ -81,7 +80,7 @@ func resourceMarkerSettingRead(ctx context.Context, d *schema.ResourceData, meta
 		return diagFromErr(err)
 	}
 
-	var detailedErr hnyerr.DetailedError
+	var detailedErr honeycombio.DetailedError
 	markerSetting, err := client.MarkerSettings.Get(ctx, d.Get("dataset").(string), d.Id())
 	if errors.As(err, &detailedErr) {
 		if detailedErr.IsNotFound() {
