+++ honeycombio/resource_query_annotation.go
@@ -9,7 +9,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
 	honeycombio "github.com/honeycombio/terraform-provider-honeycombio/client"
-	hnyerr "github.com/honeycombio/terraform-provider-honeycombio/client/errors"
 )
 
 func newQueryAnnotation() *schema.Resource {
@@ -108,7 +107,7 @@ func resourceQueryAnnotationRead(ctx context.Context, d *schema.ResourceData, me
 	}
 	dataset := d.Get("dataset").(string)
 
-	var detailedErr hnyerr.DetailedError
+	var detailedErr honeycombio.DetailedError
 	queryAnnotation, err := client.QueryAnnotations.Get(ctx, dataset, d.Id())
 	if errors.As(err, &detailedErr) {
 		if detailedErr.IsNotFound() {
